Try memoization, first cut.
From http://t0yv0.blogspot.com/2009/01/haskell-memoization-with.html

This is clearly not a nice enough way to memoize stuff (amd I'm not
using this memoize as the author intended): the program now runs
slower, and spends 49.4% time in memoize alone.

But: translatePixel calls are a fully 96.926% *fewer* when memoized,
so this is definitely a strategy worth investigating.  Certainly
without all the unsafePerformIO/MVar stuff.